+++ LisaWaveformTools/stationary_time_source.py
@@ -29,7 +29,7 @@ class StationarySourceWaveformTime(
     """Store a binary with linearly increasing frequency and constant amplitude in the time domain."""
 
     def __init__(
-            self, params: SourceParams, nt_lim_waveform: PixelGenericRange, lc: LISAConstants, *, response_mode: int = 0, t_phase_ref: np.float64 = 0.0
+            self, params: SourceParams, nt_lim_waveform: PixelGenericRange, lc: LISAConstants, *, response_mode: int = 0, t_phase_ref: float = 0.0
     ) -> None:
         """Initialize the object."""
         self._nt_lim_waveform: PixelGenericRange = nt_lim_waveform
@@ -38,7 +38,7 @@ class StationarySourceWaveformTime(
         self._nc_waveform: int = self._lc.nc_waveform
         self._consistent_extrinsic: bool = False
         self._consistent_intrinsic: bool = False
-        self._t_phase_ref: np.float64 = t_phase_ref
+        self._t_phase_ref: float = t_phase_ref
 
         self._TTs: NDArray[np.float64] = np.float64(self._nt_lim_waveform.dx) * np.arange(
             self._nt_lim_waveform.nx_min, self._nt_lim_waveform.nx_max
@@ -195,7 +195,7 @@ class StationarySourceWaveformTime(
         return self._nc_waveform
 
     @property
-    def t_phase_ref(self) -> np.float64:
+    def t_phase_ref(self) -> float:
         """Return the time at which the phase and frequency are defined."""
         return self._t_phase_ref
